requirements.txt
Twisted==12.1.0
pyOpenSSL==0.13
pyasn1==0.1.3
pycrypto==2.6
wsgiref==0.1.2
zope.interface==4.0.1
